refactor: extract streaming helpers to reduce execution_streaming.py size
src/agency_swarm/agent/execution_stream_persistence.py extract persistence helpers

src/agency_swarm/agent/execution_stream_response.py extract StreamingRunResponse wrapper class

src/agency_swarm/agent/execution_streaming.py reduce from 846 to 341 lines via extraction

src/agency_swarm/agent/execution_guardrails.py rename _extract_guardrail_texts to extract_guardrail_texts (public)

examples/guardrails_input.py examples/guardrails_output.py replace guardrails.py with focused demos

tests/** use real SDK types instead of SimpleNamespace, optimize model costs (gpt-4o to gpt-5-mini), add preface event test

.gitignore add work_context.md; AGENTS.md clarify workflow and evidence requirements
