+++ .gitignore
@@ -187,3 +187,5 @@ debug_agency_data.json
 
 .taskmaster/
 .agency_swarm_chats/
+
+work_context.md
